Fedora30: xrdp 0.9.11 workaround

+++ install_esm_fedora30.sh
@@ -1,7 +1,9 @@
 # Install Hyper-V Enhanced Session Mode on Fedora 30
 
 # Load the Hyper-V kernel module
-echo "hv_sock" | sudo tee -a /etc/modules-load.d/hv_sock.conf > /dev/null
+if ![ -f "/etc/modules-load.d/hv_sock.conf" ] || [ "$(cat /etc/modules-load.d/hv_sock.conf | grep hv_sock)" = ""  ]; then
+  echo "hv_sock" | sudo tee -a /etc/modules-load.d/hv_sock.conf > /dev/null
+fi
 
 # Configure SELinux
 # ATTENTION: This makes your system much more insecure!
@@ -10,7 +12,7 @@ sudo sed -i 's/SELINUX=enforcing/SELINUX=permissive/g' /etc/selinux/config
 # Recompile XRDP with Hyper-V enabled
 sudo dnf install -y rpmdevtools rpm-build
 rpmdev-setuptree
-dnf download --source xrdp
+dnf download --source xrdp-0.9.9-2.fc30
 rpm -ivh xrdp*.src.rpm
 sudo dnf builddep -y xrdp
 sed -i '/^%configure/ s/$/ --enable-vsock/' ~/rpmbuild/SPECS/xrdp.spec
@@ -31,10 +33,14 @@ sudo sed -i "/^bitmap_compression=.*/c\bitmap_compression=false" /etc/xrdp/xrdp.
 sudo sed -i "/^max_bpp=.*/c\max_bpp=24" /etc/xrdp/xrdp.ini
 
 sudo sed -i "/^X11DisplayOffset=.*/c\X11DisplayOffset=0" /etc/xrdp/sesman.ini
-echo "allowed_users=anybody" | sudo tee -a /etc/X11/Xwrapper.config > /dev/null
+if [ "$(cat /etc/X11/Xwrapper.config | grep allowed_users=anybody)" = "" ]; then
+  echo "allowed_users=anybody" | sudo tee -a /etc/X11/Xwrapper.config > /dev/null
+fi
 
 # Prevent dnf from reinstalling or upgrading xrdp to a version without Hyper-V support
-echo "exclude=xrdp" | sudo tee -a /etc/dnf/dnf.conf > /dev/null
+if [ "$(cat /etc/dnf/dnf.conf | grep -Pe "^exclude=xrdp$")" = "" ]; then
+  echo "exclude=xrdp" | sudo tee -a /etc/dnf/dnf.conf > /dev/null
+fi
